requirements.txt
numpy
pandas
rasterio
geopandas
shapely
scipy
networkx
matplotlib
folium
pytest
pyyaml
